modules: ramdisk read bugfix

+++ src/modules/ramdisk.c
@@ -115,7 +115,7 @@ size_t ramdiskRead(void *buffer, const char *name, size_t n) {
     size_t size = parseOctal(metadata->size);
     if(n > size) n = size;
 
-    uint8_t *data = (uint8_t *)(metadata + 512);
+    uint8_t *data = (uint8_t *)((uintptr_t)metadata + 512);
     memcpy(buffer, data, n);
     return n;
 }
